rsx: Surface cache bug fixes - Properly handle data 'transfer' when recycling frame buffer images - Clear 'recycled' surfaces before use

+++ rpcs3/Emu/RSX/D3D12/D3D12Buffer.cpp
@@ -223,9 +223,8 @@ namespace
 		const std::vector<std::pair<u32, u32>>& vertex_ranges, d3d12_data_heap& m_buffer_data)
 	{
 		size_t index_count = std::accumulate(
-			vertex_ranges.begin(), vertex_ranges.end(), 0, [](size_t acc, const auto& pair) {
-				return acc + get_index_count(
-								 rsx::method_registers.current_draw_clause.primitive, pair.second);
+			vertex_ranges.begin(), vertex_ranges.end(), 0ll, [](size_t acc, const auto& pair) {
+				return acc + get_index_count(rsx::method_registers.current_draw_clause.primitive, pair.second);
 			});
 
 		// Alloc
@@ -236,7 +235,7 @@ namespace
 		void* mapped_buffer =
 			m_buffer_data.map<void>(CD3DX12_RANGE(heap_offset, heap_offset + buffer_size));
 
-		size_t vertex_count = 0;
+		u32 vertex_count = 0;
 		for (const auto& pair : vertex_ranges)
 			vertex_count += pair.second;
